orientation-event/DeviceMotion/t026.html

<!DOCTYPE html>
<html>

  <head>
	<title>The accelerationIncludingGravity property provide the acceleration plus the acceleration equal and opposite to the gravity:</title>
	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	<script src="http://w3c-test.org/resources/testharness.js"></script>
	<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
	<script src="http://w3c-test.org/resources/WebIDLParser.js"></script>
	<script src="http://w3c-test.org/resources/idlharness.js"></script>
  <link rel='author' title='Mosquito FP7'/>
  </head>

  <body>

  
  	<h1>The accelerationIncludingGravity property provide the acceleration plus the acceleration equal and opposite to the gravity:</h1>
	<div id="log">
	</div>

	<script>
		setup({timeout:1500});
		window.addEventListener('devicemotion', function (e){
			var gvt = e.accelerationIncludingGravity;
			// X : angle par rapport a x
			x = gvt.x;
			tx.step(function() {assert_approx_equals(x, 0, 1)});
			tx.done();
			// Y : angle par rapport a y
			y = gvt.y;
			ty.step(function() {assert_approx_equals(y, 0, 1)});
			ty.done();
			// Z : acceleromètre (de base : -9.81)
			z = gvt.z;
			tz.step(function() {assert_approx_equals(z, -9.81, 1)});
			tz.done();
			
		}, false);
		alert("Put your device on a horizontal surface with the screen upmost.");

		tx = async_test("X angle must be set to zero");
		ty = async_test("Y angle must be set to zero");
		tz = async_test("Z angle must be set to -9.81");
	
	</script>
	
  </body>
  
</html>
